Adding message store view improvements.
git-svn-id: https://svn.wso2.org/repos/wso2/carbon/platform/trunk@180480 a5903396-d722-0410-b921-86c7d4935375

+++ dependencies/synapse/2.1.1-wso2v8/modules/core/src/main/java/org/apache/synapse/message/store/MessageStoreViewMBean.java
@@ -53,5 +53,5 @@ public interface MessageStoreViewMBean {
      *
      * @return the number of Messages stored in the store.
      */
-    public int getSize();
+    public long getSize();
 }
